+++ packages/react-container/src/components/index.ts
@@ -3,4 +3,5 @@ export * from './ConditionalWrapper';
 export * from './LoopDetector';
 export * from './Sentry';
 export * from './TranslationWrapper';
+export * from './Intercom';
 export * from './Logout';