+++ packages/react-container/src/index.ts
@@ -1,4 +1,11 @@
 export * from './auth';
+export type {
+  ApiKeyAuthResult,
+  OAuthAuthResult,
+  AuthResult,
+  DecodedIdToken,
+  AuthHeaders,
+} from './auth';
 export * from './utils';
 export * from './providers';
 export * from './components';