+++ packages/react-document-table/src/index.ts
@@ -1,4 +1,13 @@
 export { DocumentTable } from './DocumentTable';
 export * from './utils';
-export * from './types';
+export type {
+  CategoryByDocumentId,
+  Document,
+  DocumentsByCategory,
+  DocumentType,
+  JsonDocTypes,
+  Metadata,
+  Priority,
+  TextContainerProps,
+} from './types';
 export * from './docTypes';