+++ packages/auth-utils/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.9 Jun 9 2021
+
+- Add return type to `getProjects`
+
+## 1.0.8 Jun 9 2021
+
+- Set `project` after FakeIdp login
+
 ## 1.0.7 Jun 9 2021
 
 - Migrate to Bazel setup